requirements.txt
-i https://pypi.org/simple
sqlitedict>=1.6.0
discord-py>=1.3.4
emojis>=0.6.0